+++ R/pmap.R
@@ -10,6 +10,9 @@
 #'   arguments that `.f` will be called with. Arguments will be supply by
 #'   position if unnamed, and by name if named.
 #'
+#'   Vectors of length 1 will be recycled to any length; all other elements
+#'   must be have the same length.
+#'
 #'   A data frame is an important special case of `.l`. It will cause `.f`
 #'   to be called once for each row.
 #' @inheritParams map